Adjusted folder structure; started process of moving api call to serverless function
