+++ index.html
@@ -6,12 +6,12 @@
   <title>GitHub User Search</title>
   <meta name="description" content="">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" type="text/css" href="styles.css">
+  <link rel="stylesheet" type="text/css" href="src/styles.css">
 </head>
 
 <body>
   <div id="app"></div>
-  <script src="index.js"></script>
+  <script src="src/index.js"></script>
 </body>
 
 </html>
