+++ package.json
@@ -4,6 +4,8 @@
   "description": "Code Exercise for This Dot Labs JS Interview",
   "main": "index.html",
   "dependencies": {
+    "netlify-lambda": "^2.0.4",
+    "node-fetch": "^2.6.1",
     "vue": "^2.6.12",
     "vue-hot-reload-api": "^2.3.4"
   },
